キーバインドのキー入力のテキストフィールド用のビューを追加
