+++ macSKK/ja.lproj/Localizable.strings
@@ -76,6 +76,8 @@
 "Duplicate" = "複製";
 "Name" = "名前";
 "Are you sure you want to delete %@?" = "%@を本当に削除しますか?";
+"Modifier Flags (Required)" = "修飾キー (必須)";
+"Modifier Flags (Optional)" = "修飾キー (任意)";
 
 "KeyBindingActionHiragana" = "ひらがなモード";
 "KeyBindingActionTogglekana" = "かなカナ切り替え";
